requirements.txt
gradio==4.10.0
pillow
